Move selected filter into UI reducer
