+++ src/app/store/filters/filters.action.ts
@@ -1,10 +1,8 @@
 import { Action } from '@ngrx/store';
-import { FilterType } from '../../models/';
 
 export enum FiltersActionType {
   ADD_SELECTED_PLATFORM = '[Filters-Platform] Add Platform Filter',
   REMOVE_SELECTED_PLATFORM = '[Filters-Platform] Remove Platform Filter',
-  SET_SELECTED_FILTER = '[Filters] Set Filter'
 }
 
 export class AddSelectedPlatform implements Action {
@@ -19,13 +17,6 @@ export class RemoveSelectedPlatform implements Action {
   constructor(public payload: string) {}
 }
 
-export class SetSelectedFilter implements Action {
-  public readonly type = FiltersActionType.SET_SELECTED_FILTER;
-
-  constructor(public payload: FilterType) {}
-}
-
 export type FiltersActions =
   | AddSelectedPlatform
-  | RemoveSelectedPlatform
-  | SetSelectedFilter;
+  | RemoveSelectedPlatform;
