+++ src/app/store/ui/ui.action.ts
@@ -1,9 +1,13 @@
 import { Action } from '@ngrx/store';
 
+import { FilterType } from '../../models/';
+
 export enum UIActionType {
   TOGGLE_FILTERS_MENU = '[UI] Toggle Filters Menu',
   CLOSE_FILTERS_MENU = '[UI] Close Filters Menu',
   OPEN_FILTERS_MENU = '[UI] Open Filters Menu',
+
+  SET_SELECTED_FILTER = '[UI] Set Selected Filter'
 }
 
 export class ToggleFiltersMenu implements Action {
@@ -18,7 +22,14 @@ export class OpenFiltersMenu implements Action {
   public readonly type = UIActionType.OPEN_FILTERS_MENU;
 }
 
+export class SetSelectedFilter implements Action {
+  public readonly type = UIActionType.SET_SELECTED_FILTER;
+
+  constructor(public payload: FilterType) {}
+}
+
 export type UIActions =
   | ToggleFiltersMenu
   | CloseFiltersMenu
-  | OpenFiltersMenu;
+  | OpenFiltersMenu
+  | SetSelectedFilter;
