+++ src/app/store/ui/ui.reducer.ts
@@ -1,18 +1,21 @@
 import { createFeatureSelector, createSelector } from '@ngrx/store';
 
 import { UIActionType, UIActions } from './ui.action';
+import { FilterType } from '../../models/';
 
 /* State */
 
 export interface UIState {
   filtersMenu: {
     isOpen: boolean
+    selected: FilterType | undefined;
   };
 }
 
 const initState: UIState = {
   filtersMenu: {
-    isOpen: true
+    isOpen: true,
+    selected: FilterType.OTHER,
   }
 };
 
@@ -51,6 +54,20 @@ export function uiReducer(state = initState, action: UIActions): UIState {
       };
     }
 
+    case UIActionType.SET_SELECTED_FILTER: {
+      const selected = (state.filtersMenu.selected !== action.payload) ?
+        action.payload :
+        undefined;
+
+      return {
+        ...state,
+        filtersMenu: {
+          ...state.filtersMenu,
+          selected
+        }
+      };
+    }
+
     default: {
       return state;
     }
@@ -66,8 +83,12 @@ export const getFiltersMenuState = createSelector(
   (state: UIState) => state.filtersMenu
 );
 
+export const getSelectedFilter = createSelector(
+  getFiltersMenuState,
+  state => state.selected
+);
+
 export const getIsFiltersMenuOpen = createSelector(
   getFiltersMenuState,
   state => state.isOpen
 );
-
